Refactor controllers; remove api subdirectory
