--- app/controllers/api/registrations_controller.rb
@@ -1,6 +1,0 @@
-module Api
-  class RegistrationsController < Devise::RegistrationsController
-    clear_respond_to
-    respond_to :json
-  end
-end
