+++ app/controllers/servers_controller.rb
@@ -17,6 +17,16 @@ class ServersController < ApplicationController
     redirect_to action: "show", id: server.id
   end
 
+  def update
+    server = Server.find(params[:id])
+    server.update(server_params)
+    if server.save
+      render json: {server: server}
+    else
+      render json: {}, status: 422
+    end
+  end
+
   def oauth_redirect
     server = Server.where(state: params[:state]).first
     if server
@@ -35,5 +45,44 @@ class ServersController < ApplicationController
     else
       render status: 500, text: 'State not found'
     end
+  end
+
+  def conformance
+    server = Server.find(params[:server_id])
+    render json: {conformance: server.load_conformance(params[:refresh])}
+  end
+
+  def summary
+    summary = Server.find(params[:server_id]).summary
+    render json: {summary: summary}
+  end
+
+  def aggregate_run
+    server = Server.find(params[:server_id])
+    aggregate_run = server.aggregate_run
+    return unless aggregate_run
+    if (params[:only_failures])
+      aggregate_run.results.select! {|r| r if r['status'] != 'pass'}
+    end
+    render json: server.aggregate_run
+  end
+
+  def oauth_params
+    server = Server.find(params[:server_id])
+    if server
+      server.client_id = params[:client_id]
+      server.client_secret = params[:client_secret]
+      server.state = params[:state]
+      server.authorize_url = params[:authorize_url]
+      server.token_url = params[:token_url]
+      server.save
+      render json: { success: true }
+    end
+  end
+
+  private
+
+  def server_params
+    params.require(:server).permit(:url, :name)
   end
 end
