+++ app/controllers/tests_controller.rb
@@ -12,28 +12,4 @@ class TestsController < ApplicationController
 
     render json:{tests: @tests}
   end
-
-  def execute
-    mtest = Test.find(params[:test_id])
-
-    #todo check oath code
-    # if server.oauth_code
-    #   client1.client = server.get_oauth2_client
-    #   client1.use_oauth2_auth = true
-    # end
-
-    result = nil
-    success =  test_run.execute(mtest) do |r|
-      result = r
-    end
-
-    if success
-      render json: { tests: result.result }
-    else 
-      render json: {}
-    end
-
-  end
-
-  private
 end
